Pulled PackageSpec into a util file We're going to need this for add/remove
